requirements.txt
google-api-python-client>=2.187.0
google-auth>=2.45.0
google-auth-httplib2>=0.3.0
googleapis-common-protos>=1.72.0
anthropic>=0.75.0
